chatbot/backend/pyproject.toml
[tool.poetry]
name = "chat-bot-langchain"
version = "0.1.2"
description = "Simple chat bot to demonstrate Google cloud capabilities"
authors = ["Julien MIQUEL <julienmiquel@google.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.scripts]
start = "app.server:start"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
rag-google-cloud-vertexai-search = {path = "packages/rag-google-cloud-vertexai-search", develop = true}
google-cloud-discoveryengine = "^0.11.5"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
